chore: address review feedback for issue 180
- promote descriptor service into its own file
- keep descriptor generators grouped by type
- tighten normalization tests to assert full descriptors
