+++ internal/infrastructure/api/bitcoin/btc/descriptor_service.go
@@ -1,0 +1,79 @@
+package btc
+
+import (
+	"errors"
+	"fmt"
+	"strings"
+
+	"github.com/btcsuite/btcd/btcutil/hdkeychain"
+	"github.com/btcsuite/btcd/chaincfg"
+
+	domainWallet "github.com/hiromaily/go-crypto-wallet/internal/domain/wallet"
+)
+
+const (
+	descriptorFormatPKH    = "pkh([%s%s]%s/%d/*)"
+	descriptorFormatSHWPKH = "sh(wpkh([%s%s]%s/%d/*))"
+	descriptorFormatWPKH   = "wpkh([%s%s]%s/%d/*)"
+	descriptorFormatTR     = "tr([%s%s]%s/%d/*)"
+)
+
+// DescriptorService generates descriptors for supported single-signature templates.
+type DescriptorService struct {
+	chainParams *chaincfg.Params
+}
+
+// NewDescriptorService creates a descriptor service for the given Bitcoin network parameters.
+func NewDescriptorService(chainParams *chaincfg.Params) *DescriptorService {
+	return &DescriptorService{
+		chainParams: chainParams,
+	}
+}
+
+func (d *DescriptorService) formatDescriptor(
+	template string,
+	fingerprint string,
+	derivationPath string,
+	xpub *hdkeychain.ExtendedKey,
+	isChange bool,
+) (string, error) {
+	if xpub == nil {
+		return "", errors.New("extended public key is nil")
+	}
+
+	if xpub.IsPrivate() {
+		return "", errors.New("extended key must be public")
+	}
+
+	if d.chainParams != nil && !xpub.IsForNet(d.chainParams) {
+		return "", fmt.Errorf("extended public key network mismatch: expected %s", d.chainParams.Name)
+	}
+
+	normalizedFingerprint := strings.ToLower(strings.TrimSpace(fingerprint))
+	normalizedPath := normalizeDerivationPath(derivationPath)
+
+	if err := domainWallet.ValidateFingerprint(normalizedFingerprint); err != nil {
+		return "", fmt.Errorf("invalid fingerprint: %w", err)
+	}
+
+	if err := domainWallet.ValidateDerivationPath(normalizedPath); err != nil {
+		return "", fmt.Errorf("invalid derivation path: %w", err)
+	}
+
+	changeIndex := 0
+	if isChange {
+		changeIndex = 1
+	}
+
+	return fmt.Sprintf(template, normalizedFingerprint, normalizedPath, xpub.String(), changeIndex), nil
+}
+
+func normalizeDerivationPath(path string) string {
+	trimmed := strings.TrimSpace(path)
+
+	if strings.HasPrefix(trimmed, "m/") || strings.HasPrefix(trimmed, "M/") {
+		trimmed = trimmed[1:]
+	}
+
+	return trimmed
+}
